barber-system/requirements.txt
altair==5.5.0
attrs==25.3.0
beautiful-date==2.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
gcsa==2.6.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
httplib2==0.30.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.3.0
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
